Fix the issue that some facility page has ghost pagination
Details: In some facility page, though there is no doctors neither pagination shown, however, there is sill ul with pagination css class (which is used to extract pagination in our code). So add the code to check if there are doctors shown on the page before go check the pagination.